Move export logic in src/lib
